requirements.txt
diskcache==5.6.3
Jinja2==3.1.6
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
numpy==2.2.5
typing_extensions==4.13.2
